provider:dns: move SupportedProvider func to separate file
